Sixth Homework "Help from a friend" #1 / "Помощь друга" #1
PART 4

Implement main map pin and coordinate setting

Added a main pin marker to the map at the coordinates representing Tokyo, and made the marker draggable. The #address field is now read-only (the value was sent to the server in the form) and is automatically populated with the current coordinates of the main pin marker whenever it is moved. This change enhances UX by allowing users to visually select a location on the map.